refactor グループ作成ダイアログを切り出し
